Session auth

+++ app/vacancy_api/main.py
@@ -1,12 +1,16 @@
-from ast import dump
+from os import stat
+from fastapi.exceptions import HTTPException
+from starlette import status
 from app import vacancy_api
 from json import dumps
 from typing import Union, List, Optional
 
-from fastapi import APIRouter, Response
+from fastapi import APIRouter, Response, Cookie
 from pydantic import BaseModel
-from app.db import vacancy
 
+from app.auth.main import verify_cookie
+from app.db import vacancy
+from app.db.firm import get_firm_by_user_id
 from app.db.vacancy import (
     create_vacancy,
     delete_vacancy,
@@ -35,7 +39,11 @@ class VacancyModel(BaseModel):
 
 
 @router.post("/create_vacancy_db")
-async def create_vacancy_db(vacancy: VacancyModel):
+async def create_vacancy_db(vacancy: VacancyModel, session_cookie: str = Cookie(None)):
+    username, session_id = session_cookie.split(":")
+    session_user = await verify_cookie(username, session_id)
+    if not session_user.role == "hr":
+        raise HTTPException(status_code=403)
     await create_vacancy(
         vacancy.user_identifier,
         vacancy.title,
@@ -56,13 +64,19 @@ async def create_vacancy_db(vacancy: VacancyModel):
 
 
 @router.get("/delete_vacancy_db")
-async def delete_vacancy_db(vacancy_id: int):
+async def delete_vacancy_db(vacancy_id: int, session_cookie: str = Cookie(None)):
+    username, session_id = session_cookie.split(":")
+    user_id = await verify_cookie(username, session_id).id
+    if not vacancy_id == get_vacancies_by_user_id(user_id).id:
+        raise HTTPException(status_code=403)
     await delete_vacancy(vacancy_id)
     return Response(content="OK", media_type="plain/text", status_code=200)
 
 
 @router.get("/get_by_user_id")
-async def get_by_user_id(user_identifier: Union[int, str]):
+async def get_by_user_id(user_identifier: Union[int, str], session_cookie: str = Cookie(None)):
+    username, session_id = session_cookie.split(":")
+    await verify_cookie(username, session_id)
     content = await get_vacancies_by_user_id(user_identifier)
     json = dumps(
         [
@@ -93,7 +107,9 @@ async def get_by_user_id(user_identifier: Union[int, str]):
 
 
 @router.get("/get_by_id")
-async def get_by_id(vacancy_id: int):
+async def get_by_id(vacancy_id: int, session_cookie: str = Cookie(None)):
+    username, session_id = session_cookie.split(":")
+    await verify_cookie(username, session_id)
     vacancy = await get_vacancy_by_id(vacancy_id)
     json = dumps(
         {
